Add support for remapping edge prediction results. (#533)
Add graphstrom.gconstruct.remap_result.py to remap edge prediction
results.
Will add support for remapping node prediction results and embeddings.


#508 


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

+++ tests/end2end-tests/graphstorm-ec/mgpu_test.sh
@@ -89,7 +89,7 @@ then
 fi
 
 echo "**************dataset: Generated multilabel MovieLens EC, do inference on saved model"
-python3 -m graphstorm.run.gs_edge_classification --inference --workspace $GS_HOME/inference_scripts/ep_infer --num-trainers $NUM_INFO_TRAINERS --num-servers 1 --num-samplers 0 --part-config /data/movielen_100k_multi_label_ec/movie-lens-100k.json --ip-config ip_list.txt --ssh-port 2222 --cf ml_ec_infer.yaml  --multilabel true --num-classes 6 --node-feat-name movie:title user:feat --use-mini-batch-infer false --save-embed-path /data/gsgnn_ec/infer-emb/ --restore-model-path /data/gsgnn_ec/epoch-$best_epoch/ --save-prediction-path /data/gsgnn_ec/prediction/ --logging-file /tmp/log.txt  --logging-level debug
+python3 -m graphstorm.run.gs_edge_classification --inference --workspace $GS_HOME/inference_scripts/ep_infer --num-trainers $NUM_INFO_TRAINERS --num-servers 1 --num-samplers 0 --part-config /data/movielen_100k_multi_label_ec/movie-lens-100k.json --ip-config ip_list.txt --ssh-port 2222 --cf ml_ec_infer.yaml  --multilabel true --num-classes 6 --node-feat-name movie:title user:feat --use-mini-batch-infer false --save-embed-path /data/gsgnn_ec/infer-emb/ --restore-model-path /data/gsgnn_ec/epoch-$best_epoch/ --save-prediction-path /data/gsgnn_ec/prediction/ --logging-file /tmp/log.txt  --logging-level debug --preserve-input True
 
 error_and_exit $?
 
@@ -129,7 +129,7 @@ then
 fi
 
 cnt=$(ls -l /data/gsgnn_ec/prediction/user_rating_movie/ | grep predict | wc -l)
-if test $cnt != $NUM_INFO_TRAINERS
+if test $cnt != $NUM_INFO_TRAINERS * 2
 then
     echo "The number of saved prediction results $cnt is not equal to the number of inferers $NUM_INFO_TRAINERS"
     exit -1
@@ -169,6 +169,13 @@ echo "**************dataset: Generated multilabel MovieLens EC, do inference on 
 python3 -m graphstorm.run.gs_edge_classification --inference --workspace $GS_HOME/inference_scripts/ep_infer --num-trainers $NUM_INFO_TRAINERS --num-servers 1 --num-samplers 0 --part-config /data/movielen_100k_ec_no_test_1p_4t/movie-lens-100k.json --ip-config ip_list.txt --ssh-port 2222 --cf ml_ec_infer.yaml  --multilabel true --num-classes 6 --node-feat-name movie:title user:feat --use-mini-batch-infer false --save-embed-path /data/gsgnn_ec/infer-emb/ --restore-model-path /data/gsgnn_ec/epoch-$best_epoch/ --save-prediction-path /data/gsgnn_ec/prediction/ --no-validation true
 
 error_and_exit $?
+
+cnt=$(ls -l /data/gsgnn_ec/prediction/user_rating_movie/ | grep parquet | wc -l)
+if test $cnt != $NUM_INFO_TRAINERS
+then
+    echo "The number of remapped prediction results $cnt is not equal to the number of inferers $NUM_INFO_TRAINERS"
+    exit -1
+fi
 
 rm -fr /data/gsgnn_ec/*
